backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.21
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.2.1
python-decouple==3.8
qrcode==8.2
sqlparse==0.5.3
typing_extensions==4.13.2
